Suppresion de la colonne objet de la table mnemonique. Jointure avec DLS pour relier un mnemo à un plugin DLS
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@1286 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ Watchdogd/Dls/Dls_db.c
@@ -41,6 +41,8 @@
  gboolean Retirer_plugin_dlsDB( struct LOG *log, struct DB *db, struct CMD_TYPE_PLUGIN_DLS *dls )
   { gchar requete[200];
 
+    if (dls->id == 1) return(FALSE);                            /* On ne peut pas effacer le plugin n°1 ! */
+
     g_snprintf( (gchar *)requete, sizeof(requete),                                           /* Requete SQL */
                 "DELETE FROM %s WHERE id=%d",
                 NOM_TABLE_DLS, dls->id );
